Split components logic and add small refactor
